# services/backend/api-spec.yaml
openapi: 3.0.0
info:
  title: DeployStack Backend API
  description: API documentation for DeployStack Backend
  version: 0.20.5
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_session
...
paths:
...
  /api/teams/me/default:
    get:
      summary: Get current user default team
      tags:
        - Teams
      description: Retrieves the default team for the currently authenticated user.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Default team retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        description: Team ID
                      name:
                        type: string
                        description: Team name
                      slug:
                        type: string
                        description: Team slug
                      description:
                        type: string
                        nullable: true
                        description: Team description
                      owner_id:
                        type: string
                        description: Team owner ID
                      is_default:
                        type: boolean
                        description: Indicates if this is the user's default team
                      created_at:
                        type: string
                        format: date-time
                        description: Team creation date
                      updated_at:
                        type: string
                        format: date-time
                        description: Team last update date
                    required:
                      - id
                      - name
                      - slug
                      - description
                      - owner_id
                      - is_default
                      - created_at
                      - updated_at
                    additionalProperties: false
                    description: Team data
                  message:
                    type: string
                    description: Success message
                required:
                  - success
                  - data
                additionalProperties: false
                description: Default team retrieved successfully
        "401":
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful (false for errors)
                    default: false
                  error:
                    type: string
                    description: Error message
                  details:
                    type: array
                    description: Additional error details (validation errors)
                required:
                  - error
                additionalProperties: false
                description: Unauthorized - Authentication required
        "404":
          description: Not Found - No default team found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful (false for errors)
                    default: false
                  error:
                    type: string
                    description: Error message
                  details:
                    type: array
                    description: Additional error details (validation errors)
                required:
                  - error
                additionalProperties: false
                description: Not Found - No default team found
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful (false for errors)
                    default: false
                  error:
                    type: string
                    description: Error message
                  details:
                    type: array
                    description: Additional error details (validation errors)
                required:
                  - error
                additionalProperties: false
                description: Internal Server Error
  /api/teams/me:
    get:
      summary: Get current user teams
      tags:
        - Teams
      description: Retrieves all teams that the currently authenticated user belongs
        to, including their role in each team.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: User teams retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: Team ID
                        name:
                          type: string
                          description: Team name
                        slug:
                          type: string
                          description: Team slug
                        description:
                          type: string
                          nullable: true
                          description: Team description
                        owner_id:
                          type: string
                          description: Team owner ID
                        is_default:
                          type: boolean
                          description: Indicates if this is the user's default team
                        created_at:
                          type: string
                          format: date-time
                          description: Team creation date
                        updated_at:
                          type: string
                          format: date-time
                          description: Team last update date
                        role:
                          type: string
                          enum:
                            - team_admin
                            - team_user
                          description: User role in the team
                      required:
                        - id
                        - name
                        - slug
                        - description
                        - owner_id
                        - is_default
                        - created_at
                        - updated_at
                        - role
                      additionalProperties: false
                    description: Array of teams with user roles
                required:
                  - success
                  - data
                additionalProperties: false
                description: User teams retrieved successfully
        "401":
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful (false for errors)
                    default: false
                  error:
                    type: string
                    description: Error message
                  details:
                    type: array
                    description: Additional error details (validation errors)
                required:
                  - error
                additionalProperties: false
                description: Unauthorized - Authentication required
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful (false for errors)
                    default: false
                  error:
                    type: string
                    description: Error message
                  details:
                    type: array
                    description: Additional error details (validation errors)
                required:
                  - error
                additionalProperties: false
                description: Internal Server Error
...
  /api/teams:
    post:
      summary: Create new team
      tags:
        - Teams
      description: Creates a new team with the specified name and description. Users
        can create up to 3 teams maximum. The slug is automatically generated
        from the team name and made unique.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: Team name
                description:
                  type: string
                  maxLength: 500
                  description: Team description
              required:
                - name
              additionalProperties: false
        required: true
      security:
        - cookieAuth: []
      responses:
        "201":
          description: Team created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        description: Team ID
                      name:
                        type: string
                        description: Team name
                      slug:
                        type: string
                        description: Team slug
                      description:
                        type: string
                        nullable: true
                        description: Team description
                      owner_id:
                        type: string
                        description: Team owner ID
                      is_default:
                        type: boolean
                        description: Indicates if this is the user's default team
                      created_at:
                        type: string
                        format: date-time
                        description: Team creation date
                      updated_at:
                        type: string
                        format: date-time
                        description: Team last update date
                    required:
                      - id
                      - name
                      - slug
                      - description
                      - owner_id
                      - is_default
                      - created_at
                      - updated_at
                    additionalProperties: false
                    description: Team data
                  message:
                    type: string
                    description: Success message
                required:
                  - success
                  - data
                additionalProperties: false
                description: Team created successfully
        "400":
          description: Bad Request - Validation error or team limit reached
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful (false for errors)
                    default: false
                  error:
                    type: string
                    description: Error message
                  details:
                    type: array
                    description: Additional error details (validation errors)
                required:
                  - error
                additionalProperties: false
                description: Bad Request - Validation error or team limit reached
        "401":
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful (false for errors)
                    default: false
                  error:
                    type: string
                    description: Error message
                  details:
                    type: array
                    description: Additional error details (validation errors)
                required:
                  - error
                additionalProperties: false
                description: Unauthorized - Authentication required
        "403":
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful (false for errors)
                    default: false
                  error:
                    type: string
                    description: Error message
                  details:
                    type: array
                    description: Additional error details (validation errors)
                required:
                  - error
                additionalProperties: false
                description: Forbidden - Insufficient permissions
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful (false for errors)
                    default: false
                  error:
                    type: string
                    description: Error message
                  details:
                    type: array
                    description: Additional error details (validation errors)
                required:
                  - error
                additionalProperties: false
                description: Internal Server Error
...
servers:
  - url: http://localhost:3000
    description: Server